Fix #53 - Invalid size when updating bottom lines.
`adjustBottomLines()` was not checking for valid dimensions when
creating a StaticLayout, causing an InvalidParameterException to be
thrown. This is likely to be the case when when calling `setError()`
on a view before it's been laid out. This fix simply no-ops the
adjustment if the view has a zero size, and waits for the next layout
pass to trigger another attempt.

A case that's not handled here is when the view has a non-zero size,
but is still too small to show the views. For now, I've left this as a
crashing case, as I figure it's desirable for the user to know that
they've over-constrained the view, rather than it fallback to only
partially functioning.